Support configuration of automountServiceAccountToken field
The issue I am trying to work around with this is that I am using the
kubernetes provider for Terraform, which is supposed to ignore the incluster
config if you provide other config, but I think there is a bug where this
does not happen.

As far as I know Atlantis itself does not need the service account for
anything so think it is fine to allow disabling this.

Obviously it would be more ideal to fix the bug in Terraform but I have a
feeling this is actually down the version of kubernetes/client-go they are
using or something. This serves as an easy workaround in the meantime.